Assignment-1 Modifications
