+++ assignment-1/src/components/HelloWorld.js
@@ -1,0 +1,5 @@
+const HelloWorld = () => {
+    return <h1>Hello World!</h1>
+}
+
+export default HelloWorld;
