+++ assignment-1/src/components/InputText.js
@@ -1,0 +1,7 @@
+const InputText = (props) => {
+    const {placeholder, value, handleChange} = props;
+  
+    return <input type = "text" placeholder = {placeholder} value = {value} onChange = {handleChange} />
+}
+
+export default InputText;
