+++ assignment-1/src/components/Text.js
@@ -1,0 +1,5 @@
+const Text = (props) => {
+    return <h1>{props.message}</h1>
+}
+
+export default Text;
